Started configurations for PWA - React
