+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/user-event": "^13.5.0",
     "gumshoejs": "^5.1.2",
     "node-sass": "^9.0.0",
+    "pwafire": "^5.0.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0",
