+++ public/index.html
@@ -8,7 +8,7 @@
     <meta name="description" content="Pwani Innovation Week is a consolidated effort by Swahilipot Hub Foundation and its partners to help create a widespread culture of innovation and stimulate the innovation ecosystem in the coastal (pwani)region of eastern Africa." />
     <meta name="author" content="Chris Achinga - Swahilipot Hub" />
     <link rel="shortcut icon" href="./images/favicon.ico" />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/icon-256x256.png" />
     <!-- <link rel="stylesheet" href="./../public/css/theme.css"> -->
     <!-- <link rel="stylesheet" href="./../public/css/custom.css"> -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
