+++ src/App.js
@@ -1,7 +1,6 @@
-// import './App.css';
-import Home from './pages/home/Home';
+import Home from './pages/Home';
 
-function App() {
+const App = () => {
   return (
     <>
       <Home />
