+++ src/pages/Schedule.jsx
@@ -1,0 +1,8 @@
+
+const Schedule = () => {
+  return (
+    <div>Schedule</div>
+  )
+}
+
+export default Schedule
