+++ src/pages/Speakers.jsx
@@ -1,0 +1,8 @@
+
+const Speakers = () => {
+  return (
+    <div>Speakers</div>
+  )
+}
+
+export default Speakers
